% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list2keyval}
\alias{list2keyval}
\title{given a named list, this give a string with comma-separated keyval pairs}
\usage{
list2keyval(z)
}
\arguments{
\item{z}{the named list}
}
\description{
given a named list, this give a string with comma-separated keyval pairs
}
\examples{
list2keyval(list(color="red", width="fat", number=2))
}

