% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ped2dot}
\alias{ped2dot}
\title{write a dot file from a pedigree specified in a data frame}
\usage{
ped2dot(x, pa = "Pa", ma = "Ma", kid = "Kid", ObsNodes = character(0),
  ShowLabelNodes = character(0), ObsStyle = list(style = "filled", fillcolor
  = "gray"), outf = "pedvis-ped")
}
\arguments{
\item{x}{the data frame holding the pedigree.  There must be at
least three columns (kid, pa, and ma) which should be character
vectors, NOT factors and NOT integers.  Coerce them to characters
if they are actually integers.}

\item{pa}{string giving the name of the column holding the fathers
in each row of the pedigree.  (Or can be thought of as "parent 1" in the
case in which sex is not treated.) Default is "Pa"}

\item{ma}{the name of the column holding the mother in each row (or "parent 2"). Default is "Ma"}

\item{kid}{the name of the column the kid}

\item{ObsNodes}{a character vector of the identifiers of the kids, mas, and pas that have observed
genotype data associated with them.}

\item{ShowLabelNodes}{a character vector of the identifiers of the kids, mas, and pas that should have their
labels printed inside their nodes.}

\item{ObsStyle}{a named list of the extra style values that you want the observed nodes to have}

\item{outf}{Name of the output file (so files will be outf.dot, outf.ps, outf.pdf, etc.)}
}
\description{
write a dot file from a pedigree specified in a data frame
}

