% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simple_test_ped}
\alias{simple_test_ped}
\title{A simple pedigree to use for testing and demonstration}
\usage{
simple_test_ped()
}
\arguments{
\item{No}{parameters for now.}
}
\description{
This returns a data frame suitable for passing to ped2dot
}

